Created a word_model server
Created an asynchronous model that dispatches requests and created
unique models for each connection. The model is currently only the
simple model. Training is done when 10 packets have been collected,
which corresponds to the number of tokens. The stdin_tcp module is
capable of sending std::cin char by char, which is quite a nice
development. There still is a problem though with the backspace
character which I'll need to solve one day.